Replace code that skirted strict aliasing rules.
Instead of raw memory (array of char), allocate an array of unions of
tasks and empty structs. The empty struct ensures that construction of
the array does not perform initialization. As always, access to the
array must be through `place_task()` and `remove_task()`, to ensure
proper calls to placement new and destructors.