Reduce branching & inefficient integer instructions
In work-stealing code, reduces amount of branching by optimizing comparisons that are true iff (n1 > n2) || (n1 == 0), where n1 and n2 are unsigned and of the same type. Such comparisons are logically equivalent to (n1 - 1 >= n2). Note that Clang will optimize this automatically, but GCC up to 9.1 will not.

Moreover, there had been an integer division that could not be reduced to a right shift. That has been replaced.